requirements.txt
flask
gunicorn
sqlalchemy
flask_sqlalchemy
jinja2
python-dotenv
psycopg2-binary
